Prefer $XDG_CONFIG_HOME/electron-flags.conf over patches for each package (#1267)
This PR aims to enable the official binary cache.
This method worked for `signal-desktop` and `shogihome`, so I'm applying it here.
I'm not sure why this Arch Linux-based approach also works on NixOS, but it's definitely a useful trick.

Note: This patch will temporarily downgrade `shogime` to version 1.24.3.
However, it will be immediately updated when `flake.lock` is next refreshed. (It is already in nixos-unstable)